+++ chat/src/main/java/chat21/android/contacts/adapters/ViewHolder.java
@@ -6,10 +6,10 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.engine.DiskCacheStrategy;
 
 import chat21.android.R;
 import chat21.android.adapters.AbstractRecyclerAdapter;
+import chat21.android.contacts.activites.ContactListActivity;
 import chat21.android.user.models.IChatUser;
 import chat21.android.utils.glide.CropCircleTransformation;
 
@@ -43,11 +43,9 @@ public class ViewHolder extends RecyclerView.ViewHolder {
     private void loadContactProfileImage(IChatUser contact) {
 
         Glide.with(itemView.getContext())
-                .load(contact.getProfilePictureUrl())
-                .placeholder(R.drawable.ic_person_circle_placeholder_gray_24dp)
+                .load("")
+                .placeholder(R.drawable.ic_person_avatar)
                 .bitmapTransform(new CropCircleTransformation(itemView.getContext()))
-                .skipMemoryCache(false)
-                .diskCacheStrategy(DiskCacheStrategy.ALL)
                 .into(mProfilePicture);
     }