+++ chat/src/main/java/chat21/android/groups/adapters/GroupMembersListAdapter.java
@@ -6,11 +6,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.engine.DiskCacheStrategy;
 
 import java.util.List;
 
@@ -78,13 +76,10 @@ public class GroupMembersListAdapter extends AbstractRecyclerAdapter<IChatUser,
 
         String url = contact.getProfilePictureUrl();
 
-        // load image
-        Glide.with(getContext())
+        Glide.with(holder.itemView.getContext())
                 .load(url)
-                .placeholder(R.drawable.ic_person_circle_placeholder_gray_24dp)
-                .bitmapTransform(new CropCircleTransformation(getContext()))
-                .skipMemoryCache(false)
-                .diskCacheStrategy(DiskCacheStrategy.ALL)
+                .placeholder(R.drawable.ic_person_avatar)
+                .bitmapTransform(new CropCircleTransformation(holder.itemView.getContext()))
                 .into(holder.profilePicture);
     }