+++ chat/CHANGELOG.md
@@ -1,5 +1,17 @@
 ﻿# CHANGELOG
 
+## 0.10.2
+- rimosso supporto MultiDex - 59f1003
+- fix Glide context - 13a13c9
+- fix padding message row_recipient.xml - 69f1b98
+- fix padding message row_sender.xml - 69f1b98
+- fix toolbar ghost click in MessageListActivity.class - 833e54b
+- fix conflitto multi-dipendenze in /chat/build.gradle - 59f1003
+- fix toolbar flickering in MessageListActivity.class - bb83101
+
+## 0.10.1
+- rimosso supporto vectorDrawables (crash placeholder glide su pre-lollipop)
+
 ## 0.10
 - cambiato package da it.smart21.android.xxx a chat21.android.xxx
 - corretto errore conversWith